requirements.txt
kagglehub>=0.1.0
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0